refactor: backend api platform app factory
Introduce canonical src.api.app:create_app, domain routers/services, standardized error envelope, and baseline OpenAPI contract.
