+++ api_server.py
@@ -1,2230 +1,50 @@
 #!/usr/bin/env python3
 """
-FastAPI Web Server for the Interactive Story Engine.
+StoryForge API server entrypoint (compat wrapper).
 
-This module implements a FastAPI web server that provides RESTful API endpoints
-for the story generation system.
+Canonical app factory lives at `src.api.app:create_app`.
+This module preserves the historical `api_server:app` import path for tests,
+scripts, and Uvicorn runners.
 """
 
-import asyncio
-import html
-import json
+from __future__ import annotations
+
 import logging
 import os
-import re
-import secrets
-import threading
-import time
-import uuid
-from pathlib import Path
-from concurrent.futures import ThreadPoolExecutor
-from contextlib import asynccontextmanager
-from datetime import datetime, UTC, timedelta
-from typing import Any, Dict, List, Optional, Set
-from types import SimpleNamespace
-
-import jwt  # For token validation
+from typing import Optional
 
 import uvicorn
-from fastapi.security import APIKeyCookie
-from src.config.character_factory import CharacterFactory
-from src.agents.chronicler_agent import ChroniclerAgent
-from config_loader import get_config
-from src.agents.director_agent import DirectorAgent
-from fastapi import Depends, FastAPI, HTTPException, Request, Response, UploadFile, File, Form
-from fastapi.middleware.cors import CORSMiddleware
-from fastapi.middleware.gzip import GZipMiddleware
-from fastapi.responses import JSONResponse, StreamingResponse
-from pydantic import BaseModel, Field
-from starlette.exceptions import HTTPException as StarletteHTTPException
 
-
-from src.event_bus import EventBus
-from src.metrics.global_metrics import metrics as global_metrics
-from src.caching.registry import invalidate_by_tags
-from src.caching.global_chunk_cache import chunk_cache
+from src.api.app import create_app as _create_app
+from src.api.settings import APISettings
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
-# ===================================================================
-# Cookie Configuration for httpOnly Token Storage [SEC-001]
-# ===================================================================
-COOKIE_NAME = "access_token"
-REFRESH_COOKIE_NAME = "refresh_token"
-CSRF_COOKIE_NAME = "csrf_token"
+# Settings/constants kept for backward compatibility with tests.
+_SETTINGS = APISettings.from_env()
+JWT_SECRET_KEY = _SETTINGS.jwt_secret_key
+JWT_ALGORITHM = _SETTINGS.jwt_algorithm
 
-# Environment-based security settings
-COOKIE_SECURE = os.getenv("COOKIE_SECURE", "true").lower() in ("true", "1", "yes")  # HTTPS only in production
-COOKIE_HTTPONLY = True  # Protect against XSS
-COOKIE_SAMESITE = "lax"  # CSRF protection while allowing normal navigation
-COOKIE_MAX_AGE = 3600 * 24  # 24 hours for access token
-REFRESH_COOKIE_MAX_AGE = 3600 * 24 * 30  # 30 days for refresh token
-CSRF_TOKEN_MAX_AGE = 3600 * 24  # 24 hours for CSRF token
-
-# JWT Configuration
-JWT_SECRET_KEY = os.getenv("JWT_SECRET_KEY", os.getenv("SECRET_KEY", "development-secret-key-change-in-production"))
-JWT_ALGORITHM = "HS256"
-JWT_ACCESS_TOKEN_EXPIRE_MINUTES = 15  # Short-lived access tokens
-
-# Import World Context API Router
+# Import World Context API Router (for tests that introspect availability).
 try:
-    from apps.api.http import world_router
+    from apps.api.http import world_router as _world_router  # noqa: F401
 
     WORLD_ROUTER_AVAILABLE = True
-    logger.info("World context router successfully imported.")
-except ImportError as e:
+except ImportError as exc:
     WORLD_ROUTER_AVAILABLE = False
-    logger.warning(f"World context router not available: {e}")
+    logger.warning("World context router not available: %s", exc)
 
-try:
-    from src.shared_types import (
-        CharacterData,
-        SystemStatus,
-    )
 
-    # Bind to throwaway tuple to avoid lint "unused import" noise.
-    _SHARED_TYPE_SENTINEL = (CharacterData, SystemStatus)
-    SHARED_TYPES_AVAILABLE = True
-    logger.info("Shared types successfully imported.")
-except ImportError as e:
-    SHARED_TYPES_AVAILABLE = False
-    logger.warning(f"Shared types not available: {e}")
+def create_app():
+    return _create_app(settings=_SETTINGS)
 
 
-def _find_project_root(start_path: str) -> str:
-    """Finds the project root directory by looking for marker files."""
-    markers = ["persona_agent.py", "director_agent.py", "configs/", ".git"]
-    current_path = os.path.abspath(start_path)
-    while current_path != os.path.dirname(current_path):
-        if any(
-            os.path.exists(os.path.join(current_path, marker)) for marker in markers
-        ):
-            return current_path
-        current_path = os.path.dirname(current_path)
-    return os.path.abspath(os.getcwd())
+app = create_app()
 
 
-def _get_characters_directory_path() -> str:
-    """Gets the absolute path to the characters directory."""
-    base_character_path = "characters"
-    if not os.path.isabs(base_character_path):
-        project_root = _find_project_root(os.getcwd())
-        return os.path.join(project_root, base_character_path)
-    return base_character_path
-
-
-DEFAULT_CHARACTERS_PATH = os.path.abspath(_get_characters_directory_path())
-
-_CHARACTER_ID_SLUG_RE = re.compile(r"[^a-z0-9_-]+")
-_CHARACTER_DIRNAME_RE = re.compile(r"^[a-zA-Z0-9_-]+$")
-
-
-def _normalize_character_id(value: str) -> str:
-    raw_value = (value or "").strip().replace("\\", "/")
-    raw_value = os.path.basename(raw_value)
-    normalized = _CHARACTER_ID_SLUG_RE.sub("_", raw_value.lower()).strip("_")
-    if not normalized or normalized in {".", ".."}:
-        raise HTTPException(status_code=400, detail="Invalid character name")
-    return normalized
-
-def _structured_defaults(
-    name: str,
-    specialization: str = "Unknown",
-    faction: str = "Independent",
-) -> Dict[str, Any]:
-    return {
-        "stats": {
-            "character": {
-                "name": name,
-                "specialization": specialization,
-                "faction": faction,
-            }
-        },
-        "combat_stats": {},
-        "equipment": {"items": []},
-    }
-
-
-from src.core.service_container import ServiceContainer, get_service_container
-from src.core.system_orchestrator import SystemOrchestrator, OrchestratorConfig
-from src.services.api_service import ApiOrchestrationService
-from src.api.schemas import (
-    HealthResponse,
-    ErrorResponse,
-    CharactersListResponse,
-    SimulationRequest,
-    SimulationResponse,
-    CharacterDetailResponse,
-    FileCount,
-    CampaignsListResponse,
-    CampaignCreationRequest,
-    CampaignCreationResponse,
-    LoginRequest,
-    AuthResponse,
-    RefreshTokenRequest,
-    CSRFTokenResponse,
-    InvalidationRequest,
-    ChunkInRequest,
-    GuestSessionResponse,
-    WorkspaceCharacterCreateRequest,
-    WorkspaceCharacterUpdateRequest,
-)
-
-from src.workspaces import (
-    FilesystemCharacterStore,
-    FilesystemWorkspaceStore,
-    GuestSessionManager,
-)
-
-@asynccontextmanager
-async def lifespan(app: FastAPI):
-    """FastAPI lifespan event handler."""
-    logger.info("Starting StoryForge AI API server...")
-    try:
-        # Validate configuration during startup
-        config = get_config()
-        logger.info("Configuration loaded successfully")
-
-        # Initialize Service Container
-        container = get_service_container()
-
-        # Guest workspaces (filesystem-backed)
-        if getattr(app.state, "workspace_store", None) is None:
-            try:
-                workspace_root = Path(_find_project_root(os.getcwd())) / "data" / "workspaces"
-                workspace_store = FilesystemWorkspaceStore(workspace_root)
-                character_store = FilesystemCharacterStore(workspace_store)
-                guest_session_manager = GuestSessionManager(
-                    workspace_store,
-                    secret_key=JWT_SECRET_KEY,
-                    algorithm=JWT_ALGORITHM,
-                )
-                app.state.workspace_store = workspace_store
-                app.state.workspace_character_store = character_store
-                app.state.guest_session_manager = guest_session_manager
-                logger.info("Guest workspace store initialized")
-            except Exception as ws_error:
-                logger.error(f"Failed to initialize guest workspace store: {ws_error}")
-                app.state.workspace_store = None
-                app.state.workspace_character_store = None
-                app.state.guest_session_manager = None
-        else:
-            logger.info("Guest workspace store preconfigured; skipping initialization")
-
-        # Setup EventBus
-        try:
-            from src.event_bus import EventBus as EventBusClass
-            global_event_bus = EventBusClass()
-            app.state.event_bus = global_event_bus
-            
-            # Register Global EventBus in Container
-            container.register_singleton(EventBusClass, global_event_bus)
-            logger.info("Global EventBus initialized and registered")
-        except Exception as bus_error:
-            logger.warning(f"Could not initialize EventBus: {bus_error}")
-            global_event_bus = None
-
-        # Initialize System Orchestrator (Functional Core)
-        try:
-            orchestrator = SystemOrchestrator(
-                event_bus=global_event_bus,
-                config=OrchestratorConfig(debug_logging=True)
-            )
-            await orchestrator.startup()
-            container.register_singleton(SystemOrchestrator, orchestrator)
-            logger.info("System Orchestrator initialized")
-        except Exception as orch_error:
-            logger.error(f"Failed to initialize SystemOrchestrator: {orch_error}")
-            # We continue, but orchestration might fail
-            orchestrator = None
-
-        # Initialize API Service (Imperative Shell Bridge)
-        if orchestrator and global_event_bus:
-            character_factory = CharacterFactory(global_event_bus)
-            api_service = ApiOrchestrationService(orchestrator, global_event_bus, character_factory)
-            container.register_singleton(ApiOrchestrationService, api_service)
-            app.state.api_service = api_service
-            logger.info("ApiOrchestrationService initialized")
-        else:
-            logger.warning("ApiOrchestrationService skipped due to missing dependencies")
-            app.state.api_service = None
-
-    except Exception as e:
-        logger.error(f"Configuration error during startup: {e}")
-        raise e
-    
-    yield
-    
-    logger.info("Shutting down StoryForge AI API server.")
-    # Shutdown services
-    try:
-        container = get_service_container()
-        # await container.shutdown_all_services() # If we fully implemented container lifecycle
-        if hasattr(app.state, 'api_service') and app.state.api_service:
-            await app.state.api_service.stop_simulation()
-    except Exception as e:
-        logger.error(f"Error during shutdown: {e}")
-
-
-API_START_TIME = datetime.now(UTC)
-
-
-def _uptime_seconds() -> float:
-    return (datetime.now(UTC) - API_START_TIME).total_seconds()
-
-
-
-
-app = FastAPI(
-    title="StoryForge AI API",
-    description="RESTful API for the StoryForge AI Interactive Story Engine.",
-    version="1.0.0",
-    lifespan=lifespan,
-)
-
-app.add_middleware(GZipMiddleware, minimum_size=1000)
-app.add_middleware(
-    CORSMiddleware,
-    allow_origins=["http://localhost:3000", "http://127.0.0.1:3000"],
-    allow_credentials=True,
-    allow_methods=["*"],
-    allow_headers=["*"],
-)
-
-@app.middleware("http")
-async def _cors_preflight_middleware(request: Request, call_next):
-    if request.method == "OPTIONS":
-        return Response(
-            status_code=204,
-            headers={
-                "Access-Control-Allow-Origin": "http://localhost:3000",
-                "Access-Control-Allow-Methods": "*",
-                "Access-Control-Allow-Headers": "*",
-                "Access-Control-Allow-Credentials": "true",
-            },
-        )
-    return await call_next(request)
-
-
-
-
-# Include Prompts Router
-try:
-    from src.api.prompts_router import router as prompts_router
-
-    app.include_router(prompts_router, prefix="/api")
-    logger.info("Prompts router included with prefix /api")
-except ImportError as e:
-    logger.warning(f"Prompts router not available: {e}")
-
-# Include Decision Router for user participatory interaction
-try:
-    from src.decision import (
-        decision_router,
-        initialize_decision_system,
-        InteractionPauseController,
-        DecisionPointDetector,
-        NegotiationEngine,
-    )
-
-    # Create decision system components
-    _decision_pause_controller = InteractionPauseController(default_timeout=120)
-    _decision_detector = DecisionPointDetector(
-        tension_threshold=7.0,
-        intensity_threshold=7.0,
-        always_detect_interval=3,  # For testing: every 3 turns
-    )
-    _decision_negotiation_engine = NegotiationEngine()
-
-    # Include router
-    app.include_router(decision_router, prefix="/api")
-    DECISION_ROUTER_AVAILABLE = True
-    logger.info("Decision router included with prefix /api")
-except ImportError as e:
-    DECISION_ROUTER_AVAILABLE = False
-    _decision_pause_controller = None
-    _decision_detector = None
-    _decision_negotiation_engine = None
-    logger.warning(f"Decision router not available: {e}")
-
-
-@app.exception_handler(StarletteHTTPException)
-async def http_exception_handler(request: Request, exc: StarletteHTTPException):
-    """Custom HTTP exception handler to format errors as expected by tests."""
-    error_map = {
-        404: "Not Found",
-        400: "Bad Request",
-        422: "Validation Error",
-        500: "Internal Server Error",
-    }
-
-    # Log errors based on status code
-    if exc.status_code == 404:
-        logger.warning(f"404 error for path {request.url.path}: {exc.detail}")
-        detail = "The requested endpoint does not exist."
-    elif exc.status_code >= 500:
-        logger.error(
-            f"Server error {exc.status_code} for path {request.url.path}: {exc.detail}"
-        )
-        detail = exc.detail
-    else:
-        detail = exc.detail
-
-    return JSONResponse(
-        status_code=exc.status_code,
-        content={
-            "error": error_map.get(exc.status_code, "Unknown Error"),
-            "detail": detail,
-        },
-    )
-
-
-@app.exception_handler(HTTPException)
-async def fastapi_exception_handler(request: Request, exc: HTTPException):
-    """Custom FastAPI exception handler to format errors as expected by tests."""
-    error_map = {
-        404: "Not Found",
-        400: "Bad Request",
-        422: "Validation Error",
-        500: "Internal Server Error",
-    }
-
-    return JSONResponse(
-        status_code=exc.status_code,
-        content={
-            "error": error_map.get(exc.status_code, "Unknown Error"),
-            "detail": exc.detail,
-        },
-    )
-
-
-
-
-@app.get("/cache/metrics")
-async def cache_metrics() -> Dict[str, Any]:
-    """Return cache/coordinator metrics snapshot (internal)."""
-    try:
-        return global_metrics.snapshot().to_dict()
-    except Exception as e:
-        logger.error(f"metrics snapshot error: {e}")
-        raise HTTPException(status_code=500, detail="metrics error")
-
-
-@app.post("/cache/invalidate")
-async def cache_invalidate(req: InvalidationRequest) -> Dict[str, Any]:
-    """Invalidate cache entries that include all provided tags (intersection)."""
-    try:
-        removed = invalidate_by_tags(req.all_of)
-        return {"removed": int(removed)}
-    except Exception as e:
-        logger.error(f"invalidation error: {e}")
-        raise HTTPException(status_code=500, detail="invalidation error")
-
-
-@app.post("/cache/chunk/{key}")
-async def cache_chunk_append(key: str, req: ChunkInRequest) -> Dict[str, Any]:
-    try:
-        chunk_cache.add_chunk(key, req.seq, req.data)
-        return {"ok": True}
-    except Exception as e:
-        logger.error(f"chunk append error: {e}")
-        raise HTTPException(status_code=500, detail="chunk append error")
-
-
-@app.post("/cache/chunk/{key}/complete")
-async def cache_chunk_complete(key: str) -> Dict[str, Any]:
-    try:
-        chunk_cache.mark_complete(key)
-        return {"ok": True}
-    except Exception as e:
-        logger.error(f"chunk complete error: {e}")
-        raise HTTPException(status_code=500, detail="chunk complete error")
-
-
-@app.get("/cache/stream/{key}")
-async def cache_stream(key: str):
-    import asyncio
-
-    async def event_gen():
-        last_seq = -1
-        idle = 0
-        while True:
-            chunks = chunk_cache.get_since(key, last_seq)
-            for c in chunks:
-                last_seq = c.seq
-                yield f"data: {c.data}\n\n"
-            if chunk_cache.is_complete(key):
-                break
-            await asyncio.sleep(0.2)
-            idle += 1
-            if idle > 300:  # ~60s timeout
-                break
-        # end of stream signal
-        yield "event: done\ndata: complete\n\n"
-
-    return StreamingResponse(event_gen(), media_type="text/event-stream")
-
-
-@app.get("/", response_model=HealthResponse)
-async def root() -> Dict[str, Any]:
-    """Provides a basic health check for the API."""
-    logger.info("Root endpoint accessed for health check")
-    response_data = {
-        "message": "StoryForge AI Interactive Story Engine is running!",
-        "status": "ok",
-        "timestamp": datetime.now(UTC).isoformat(),
-    }
-    logger.debug(f"Root endpoint response: {response_data}")
-    return response_data
-
-
-@app.get("/health")
-async def health_check() -> Dict[str, Any]:
-    """Comprehensive health check endpoint."""
-    import datetime
-
-    try:
-        # Test if logging.Formatter is causing issues (for test mocking)
-        import logging
-
-        logging.Formatter("%(name)s - %(levelname)s - %(message)s")
-
-        # Test configuration loading
-        get_config()
-        config_status = "loaded"
-        status = "healthy"
-        logger.info("Health check endpoint accessed")
-    except Exception as e:
-        logger.error(f"Health check configuration error: {e}")
-        # If it's a severe error (like in testing), raise HTTP 500
-        if "Severe system error" in str(e):
-            raise HTTPException(status_code=500, detail=str(e))
-        config_status = "error"
-        status = "degraded"
-
-    health_data = {
-        "status": status,
-        "api": "running",
-        "timestamp": datetime.datetime.now().isoformat(),
-        "version": "1.0.0",
-        "config": config_status,
-        "uptime": _uptime_seconds(),
-    }
-
-    logger.debug(f"Health check response: {health_data}")
-    return health_data
-
-
-@app.get("/meta/system-status")
-async def system_status() -> Dict[str, Any]:
-    return {
-        "status": "operational",
-        "uptime": _uptime_seconds(),
-        "version": "1.0.0",
-        "components": {
-            "api": "online",
-            "simulation": "idle",
-            "cache": "available",
-        },
-    }
-
-
-
-# Validated orchestration state and background runner removed in favor of ApiOrchestrationService.
-# Use app.state.api_service for relevant functionality.
-
-
-
-
-@app.get("/api/orchestration/status")
-async def get_orchestration_status() -> Dict[str, Any]:
-    """
-    Get current orchestration/pipeline status for dashboard display.
-    Delegates to ApiOrchestrationService.
-    """
-    if not hasattr(app.state, 'api_service') or not app.state.api_service:
-        return {
-            "success": False,
-            "message": "Orchestration service not available",
-            "data": {"status": "error"}
-        }
-
-    status = await app.state.api_service.get_status()
-    return {
-        "success": True,
-        "data": status
-    }
-
-
-@app.post("/api/orchestration/start")
-async def start_orchestration(request: Optional[Dict[str, Any]] = None) -> Dict[str, Any]:
-    """
-    Start real story generation orchestration pipeline.
-    delegates to ApiOrchestrationService.
-    """
-    if not hasattr(app.state, 'api_service') or not app.state.api_service:
-        raise HTTPException(status_code=503, detail="Orchestration service not initialized")
-
-    # Parse request parameters
-    params = request or {}
-    total_turns = params.get("total_turns", 3)
-    start_turn = params.get("start_turn", 1) # Deprecated kept for compat, actually unused by service logic simplified
-    character_names = params.get("character_names")
-
-    # Default character logic
-    if not character_names:
-        try:
-            characters_path = _get_characters_directory_path()
-            available_chars: Set[str] = set()
-            if hasattr(characters_path, 'iterdir') or isinstance(characters_path, str): 
-                 # Handle path string vs Path object
-                 p = Path(characters_path) if isinstance(characters_path, str) else characters_path
-                 if p.exists():
-                     for item in p.iterdir():
-                         if item.is_dir() and not item.name.startswith('.'):
-                             available_chars.add(item.name)
-            
-            character_names = sorted(list(available_chars))[:3]
-            if not character_names:
-                character_names = ["pilot", "scientist", "engineer"]
-        except Exception as e:
-            logger.warning(f"Failed to fetch characters, using defaults: {e}")
-            character_names = ["pilot", "scientist", "engineer"]
-
-    sim_request = SimulationRequest(
-        character_names=character_names,
-        turns=total_turns
-    )
-
-    try:
-        result = await app.state.api_service.start_simulation(sim_request)
-        return result
-    except ValueError as val_err:
-        return {
-            "success": False,
-            "message": str(val_err),
-            "data": await app.state.api_service.get_status()
-        }
-    except Exception as e:
-        logger.error(f"Failed to start orchestration: {e}")
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@app.post("/api/orchestration/stop")
-async def stop_orchestration() -> Dict[str, Any]:
-    """Stop orchestration pipeline gracefully."""
-    if not hasattr(app.state, 'api_service') or not app.state.api_service:
-         return {"success": False, "message": "Service unavailable"}
-    
-    return await app.state.api_service.stop_simulation()
-
-
-@app.get("/api/orchestration/narrative")
-async def get_narrative() -> Dict[str, Any]:
-    """
-    Get the generated story narrative from the last orchestration run.
-    """
-    if not hasattr(app.state, 'api_service') or not app.state.api_service:
-        return {"success": False, "data": {}}
-
-    narrative = await app.state.api_service.get_narrative()
-    return {
-        "success": True, 
-        "data": {
-            "story": narrative.get("story", ""),
-            "participants": narrative.get("participants", []),
-            "turns_completed": narrative.get("turns_completed", 0),
-            "last_generated": narrative.get("last_generated"),
-            "has_content": bool(narrative.get("story", "")),
-        }
-    }
-
-
-@app.get("/meta/policy")
-async def policy_info() -> Dict[str, Any]:
-    return {
-        "brand_status": "Generic Sci-Fi",
-        "compliance": {
-            "intellectual_property": "debranded",
-            "content_filters": "enabled",
-        },
-        "last_reviewed": datetime.now(UTC).isoformat(),
-    }
-
-@app.post("/simulations", response_model=SimulationResponse)
-async def run_simulation(sim_request: SimulationRequest) -> SimulationResponse:
-    """Legacy-compatible simulation endpoint used by tests and clients."""
-    start_time = time.time()
-
-    event_bus = EventBus()
-    character_factory = CharacterFactory(event_bus)
-    director = DirectorAgent(event_bus=event_bus)
-    chronicler = ChroniclerAgent(event_bus=event_bus, character_names=sim_request.character_names)
-
-    missing_characters: List[str] = []
-    agents = []
-    for name in sim_request.character_names:
-        try:
-            agent = character_factory.create_character(name)
-        except FileNotFoundError:
-            missing_characters.append(name)
-            continue
-        except Exception as e:
-            raise HTTPException(
-                status_code=400, detail=f"Failed to load character {name}: {e}"
-            )
-        agents.append(agent)
-        try:
-            director.register_agent(agent)
-        except Exception:
-            pass
-
-    if missing_characters:
-        raise HTTPException(
-            status_code=404,
-            detail=f"Characters not found: {', '.join(missing_characters)}",
-        )
-
-    turns = sim_request.turns or 3
-    for _ in range(turns):
-        try:
-            director.run_turn()
-        except Exception as e:
-            logger.error(f"Director turn failed: {e}")
-            continue
-
-    try:
-        story = chronicler.transcribe_log(director.campaign_log_file)
-    except Exception as e:
-        participants = ", ".join(sim_request.character_names)
-        story = (
-            f"A story featuring {participants} was generated, but detailed transcription failed: {e}"
-        )
-
-    duration = time.time() - start_time
-    return SimulationResponse(
-        story=story,
-        participants=sim_request.character_names,
-        turns_executed=turns,
-        duration_seconds=duration,
-    )
-
-
-@app.get("/characters", response_model=CharactersListResponse)
-async def get_characters() -> CharactersListResponse:
-    """Retrieves a list of available characters."""
-    try:
-        characters_path = _get_characters_directory_path()
-        logger.info(f"Looking for characters in: {characters_path}")
-
-        if not os.path.isdir(characters_path):
-            logger.warning(f"Characters directory not found at: {characters_path}")
-            os.makedirs(characters_path, exist_ok=True)
-
-        characters: Set[str] = set()
-        if os.path.isdir(characters_path):
-            for item in os.listdir(characters_path):
-                item_path = os.path.join(characters_path, item)
-                if os.path.isdir(item_path):
-                    characters.add(item)
-                    logger.debug(f"Found character directory: {item}")
-
-        characters = sorted(characters)
-        logger.info(f"Found {len(characters)} characters: {characters}")
-        return CharactersListResponse(characters=characters)
-    except FileNotFoundError as e:
-        logger.error(f"File not found error: {e}")
-        raise HTTPException(status_code=404, detail="Characters directory not found.")
-    except PermissionError as e:
-        logger.error(f"Permission error accessing characters: {e}")
-        raise HTTPException(
-            status_code=500, detail="Permission denied accessing characters directory."
-        )
-    except Exception as e:
-        logger.error(f"Unexpected error retrieving characters: {e}", exc_info=True)
-        raise HTTPException(
-            status_code=500, detail=f"Failed to retrieve characters: {str(e)}"
-        )
-
-
-
-@app.post("/characters", response_model=CharacterDetailResponse)
-async def create_character(
-    name: str = Form(...),
-    description: str = Form(...),
-    faction: Optional[str] = Form("Independent"),
-    role: Optional[str] = Form("Unknown"),
-    stats: Optional[str] = Form(None),
-    equipment: Optional[str] = Form(None),
-    files: List[UploadFile] = File(None),
-) -> CharacterDetailResponse:
-    """Creates a new character."""
-    try:
-        characters_path = _get_characters_directory_path()
-        os.makedirs(characters_path, exist_ok=True)
-
-        try:
-            stats_data = json.loads(stats) if stats else {}
-        except json.JSONDecodeError as exc:
-            raise HTTPException(status_code=400, detail=f"Invalid stats JSON: {exc}") from exc
-
-        try:
-            equipment_data = json.loads(equipment) if equipment else []
-        except json.JSONDecodeError as exc:
-            raise HTTPException(status_code=400, detail=f"Invalid equipment JSON: {exc}") from exc
-
-        character_id = uuid.uuid4().hex
-        character_path = os.path.join(characters_path, character_id)
-        while os.path.exists(character_path):
-            character_id = uuid.uuid4().hex
-            character_path = os.path.join(characters_path, character_id)
-
-        os.makedirs(character_path, exist_ok=False)
-
-        # Save basic info
-        # In a real implementation, we would use the CharacterFactory or Agent to persist this properly.
-        # For now, we'll create the directory structure and a basic profile.
-        
-        # Create character markdown
-        profile_content = f"""# {name}
-
-## Overview
-**Role:** {role}
-**Faction:** {faction}
-
-{description}
-
-## Stats
-{json.dumps(stats_data, indent=2)}
-
-## Equipment
-{json.dumps(equipment_data, indent=2)}
-"""
-        markdown_filename = f"character_{character_id}.md"
-        with open(
-            os.path.join(character_path, markdown_filename), "w", encoding="utf-8"
-        ) as f:
-            f.write(profile_content)
-
-        # Handle file uploads
-        if files:
-            uploads_dir = os.path.join(character_path, "uploads")
-            os.makedirs(uploads_dir, exist_ok=True)
-            for file in files:
-                upload_id = uuid.uuid4().hex
-                file_path = os.path.join(uploads_dir, f"upload_{upload_id}.bin")
-                with open(file_path, "wb") as f:
-                    content = await file.read()
-                    f.write(content)
-
-        logger.info("Created character")
-
-        # Return the created character details
-        # We reuse get_character_detail logic or construct response
-        return CharacterDetailResponse(
-            character_id=character_id,
-            character_name=character_id,
-            name=name,
-            background_summary=description,
-            personality_traits="Unknown", # Placeholder
-            current_status="Active",
-            narrative_context="Newly created character",
-            skills=stats_data,
-            inventory=[item.get("name", "Unknown") for item in equipment_data],
-            metadata={"role": role, "faction": faction},
-            structured_data={
-                "stats": {"character": {"name": name, "faction": faction, "specialization": role}},
-                "combat_stats": stats_data,
-                "equipment": {"items": equipment_data}
-            }
-        )
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"Failed to create character: {e}", exc_info=True)
-        raise HTTPException(status_code=500, detail=f"Failed to create character: {str(e)}")
-
-
-@app.get("/characters/{character_id}", response_model=CharacterDetailResponse)
-async def get_character_detail(character_id: str) -> CharacterDetailResponse:
-    """Retrieves detailed information about a specific character."""
-    try:
-        raw_character_id = (character_id or "").strip()
-        safe_character_id = os.path.basename(raw_character_id)
-        if (
-            not safe_character_id
-            or safe_character_id in {".", ".."}
-            or safe_character_id != raw_character_id
-            or not _CHARACTER_DIRNAME_RE.fullmatch(safe_character_id)
-        ):
-            raise HTTPException(status_code=400, detail="Invalid character_id")
-        # Load character data using CharacterFactory (filesystem access is constrained inside the factory).
-        try:
-            event_bus = EventBus()
-            character_factory = CharacterFactory(event_bus)
-            agent = character_factory.create_character(safe_character_id)
-            character = agent.character
-
-            structured = getattr(character, "structured_data", None)
-            if not structured:
-                structured = _structured_defaults(
-                    getattr(character, "name", character_id),
-                    getattr(character, "specialization", "Unknown"),
-                    getattr(character, "faction", "Independent"),
-                )
-
-            return CharacterDetailResponse(
-                character_id=safe_character_id,
-                character_name=safe_character_id,
-                name=character.name,
-                background_summary=getattr(
-                    character, "background_summary", "No background available"
-                ),
-                personality_traits=getattr(
-                    character,
-                    "personality_traits",
-                    "No personality traits available",
-                ),
-                current_status=getattr(character, "current_status", "Unknown"),
-                narrative_context=getattr(
-                    character, "narrative_context", "No narrative context"
-                ),
-                skills=getattr(character, "skills", {}),
-                relationships=getattr(character, "relationships", {}),
-                current_location=getattr(character, "current_location", "Unknown"),
-                inventory=getattr(character, "inventory", []),
-                metadata=getattr(character, "metadata", {}),
-                structured_data=structured,
-            )
-        except FileNotFoundError:
-            raise HTTPException(
-                status_code=404, detail=f"Character '{safe_character_id}' not found"
-            )
-        except Exception:
-            logger.error("Error loading character", exc_info=True)
-            structured = _structured_defaults(
-                safe_character_id.replace("_", " ").title(), "Unknown"
-            )
-            return CharacterDetailResponse(
-                character_id=safe_character_id,
-                character_name=safe_character_id,
-                name=safe_character_id.replace("_", " ").title(),
-                background_summary="Character data could not be loaded",
-                personality_traits="Unknown",
-                current_status="Data unavailable",
-                narrative_context="Character files could not be parsed",
-                structured_data=structured,
-            )
-
-    except HTTPException:
-        raise
-    except Exception:
-        logger.error("Unexpected error retrieving character", exc_info=True)
-        raise HTTPException(
-            status_code=500, detail="Failed to retrieve character details"
-        )
-
-
-@app.get("/characters/{character_id}/enhanced")
-async def get_character_enhanced(character_id: str) -> Dict[str, Any]:
-    """Retrieves enhanced character context and analysis data."""
-    try:
-        raw_character_id = (character_id or "").strip()
-        safe_character_id = os.path.basename(raw_character_id)
-        if (
-            not safe_character_id
-            or safe_character_id in {".", ".."}
-            or safe_character_id != raw_character_id
-            or not _CHARACTER_DIRNAME_RE.fullmatch(safe_character_id)
-        ):
-            raise HTTPException(status_code=400, detail="Invalid character_id")
-
-        event_bus = EventBus()
-        character_factory = CharacterFactory(event_bus)
-        try:
-            agent = character_factory.create_character(safe_character_id)
-        except FileNotFoundError:
-            raise HTTPException(status_code=404, detail="Character not found")
-
-        character_data = getattr(agent, "character_data", {}) or {}
-        hybrid = character_data.get("hybrid_context", {}) if isinstance(character_data, dict) else {}
-        enhanced_context = (
-            hybrid.get("markdown_content")
-            or getattr(agent, "character_context", "")
-            or getattr(agent.character, "narrative_context", "")
-            or ""
-        )
-
-        psychological_profile = (
-            character_data.get("psychological_profile", {})
-            if isinstance(character_data, dict)
-            else {}
-        )
-
-        tactical_analysis = {
-            "combat_stats": character_data.get("combat_stats", {})
-            if isinstance(character_data, dict)
-            else {},
-            "decision_weights": character_data.get("decision_weights", {})
-            if isinstance(character_data, dict)
-            else {},
-            "relationship_scores": character_data.get("relationship_scores", {})
-            if isinstance(character_data, dict)
-            else {},
-        }
-
-        return {
-            "character_id": safe_character_id,
-            "enhanced_context": enhanced_context,
-            "psychological_profile": psychological_profile,
-            "tactical_analysis": tactical_analysis,
-        }
-    except HTTPException:
-        raise
-    except Exception:
-        logger.error("Error retrieving enhanced character", exc_info=True)
-        raise HTTPException(
-            status_code=500,
-            detail="Failed to retrieve enhanced character details",
-        )
-
-
-@app.get("/campaigns", response_model=CampaignsListResponse)
-async def get_campaigns() -> CampaignsListResponse:
-    """Retrieves a list of available campaigns."""
-    try:
-        # Look for campaign files in logs or campaigns directory
-        campaigns_paths = ["campaigns", "logs", "private/campaigns"]
-        campaigns = []
-
-        for campaigns_path in campaigns_paths:
-            if os.path.isdir(campaigns_path):
-                for item in os.listdir(campaigns_path):
-                    if item.endswith(".md") or item.endswith(".json"):
-                        campaign_name = item.replace(".md", "").replace(".json", "")
-                        if campaign_name not in campaigns:
-                            campaigns.append(campaign_name)
-
-        campaigns = sorted(campaigns)
-        logger.info(f"Found {len(campaigns)} campaigns: {campaigns}")
-        return CampaignsListResponse(campaigns=campaigns)
-
-    except Exception as e:
-        logger.error(f"Error retrieving campaigns: {e}", exc_info=True)
-        return CampaignsListResponse(campaigns=[])
-
-
-def _get_optional_workspace_id(request: Request) -> Optional[str]:
-    manager = getattr(request.app.state, "guest_session_manager", None)
-    store = getattr(request.app.state, "workspace_store", None)
-    if not manager or not store:
-        return None
-    token = request.cookies.get(manager.cookie_name)
-    if not token:
-        return None
-    workspace_id = manager.decode(token)
-    if not workspace_id:
-        return None
-    store.get_or_create(workspace_id)
-    return workspace_id
-
-
-def _require_workspace_id(request: Request, response: Response) -> str:
-    manager = getattr(request.app.state, "guest_session_manager", None)
-    store = getattr(request.app.state, "workspace_store", None)
-    if not manager or not store:
-        raise HTTPException(status_code=503, detail="Workspace service unavailable")
-
-    token = request.cookies.get(manager.cookie_name)
-    result = manager.resolve_or_create(token)
-
-    response.set_cookie(
-        manager.cookie_name,
-        manager.encode(result.workspace_id),
-        httponly=True,
-        secure=COOKIE_SECURE,
-        samesite=COOKIE_SAMESITE,
-        max_age=manager.cookie_max_age_seconds(),
-    )
-    return result.workspace_id
-
-
-def _workspace_record_to_character_detail(record: Dict[str, Any]) -> CharacterDetailResponse:
-    char_id = str(record.get("id") or "")
-    char_name = str(record.get("name") or char_id)
-    inventory = record.get("inventory")
-    if not isinstance(inventory, list):
-        inventory = []
-    skills = record.get("skills")
-    if not isinstance(skills, dict):
-        skills = {}
-    relationships = record.get("relationships")
-    if not isinstance(relationships, dict):
-        relationships = {}
-    metadata = record.get("metadata")
-    if not isinstance(metadata, dict):
-        metadata = {}
-    structured_data = record.get("structured_data")
-    if not isinstance(structured_data, dict):
-        structured_data = {}
-
-    return CharacterDetailResponse(
-        character_id=char_id,
-        character_name=char_name,
-        name=char_name,
-        background_summary=str(record.get("background_summary") or ""),
-        personality_traits=str(record.get("personality_traits") or ""),
-        current_status=str(record.get("current_status") or "active"),
-        narrative_context=str(record.get("narrative_context") or ""),
-        skills=skills,
-        relationships=relationships,
-        current_location=str(record.get("current_location") or ""),
-        inventory=inventory,
-        metadata=metadata,
-        structured_data=structured_data,
-    )
-
-
-@app.post("/api/guest/session", response_model=GuestSessionResponse, tags=["Guest"])
-async def create_or_resume_guest_session(request: Request, response: Response) -> GuestSessionResponse:
-    """Create or resume a guest session backed by a filesystem workspace."""
-    manager = getattr(request.app.state, "guest_session_manager", None)
-    if not manager:
-        raise HTTPException(status_code=503, detail="Workspace service unavailable")
-
-    token = request.cookies.get(manager.cookie_name)
-    result = manager.resolve_or_create(token)
-    response.set_cookie(
-        manager.cookie_name,
-        manager.encode(result.workspace_id),
-        httponly=True,
-        secure=COOKIE_SECURE,
-        samesite=COOKIE_SAMESITE,
-        max_age=manager.cookie_max_age_seconds(),
-    )
-    return GuestSessionResponse(workspace_id=result.workspace_id, created=result.created)
-
-
-@app.get("/api/workspace/export", tags=["Guest"])
-async def export_workspace_zip(
-    request: Request,
-    workspace_id: str = Depends(_require_workspace_id),
-) -> StreamingResponse:
-    """Export the current workspace as a zip file."""
-    store = getattr(request.app.state, "workspace_store", None)
-    manager = getattr(request.app.state, "guest_session_manager", None)
-    if not store or not manager:
-        raise HTTPException(status_code=503, detail="Workspace service unavailable")
-    zip_bytes = store.export_zip(workspace_id)
-    filename = f"workspace-{workspace_id}.zip"
-    headers = {"Content-Disposition": f'attachment; filename="{filename}"'}
-    streaming = StreamingResponse(iter([zip_bytes]), media_type="application/zip", headers=headers)
-    streaming.set_cookie(
-        manager.cookie_name,
-        manager.encode(workspace_id),
-        httponly=True,
-        secure=COOKIE_SECURE,
-        samesite=COOKIE_SAMESITE,
-        max_age=manager.cookie_max_age_seconds(),
-    )
-    return streaming
-
-
-@app.post("/api/workspace/import", response_model=GuestSessionResponse, tags=["Guest"])
-async def import_workspace_zip(
-    request: Request,
-    response: Response,
-    archive: UploadFile = File(...),
-) -> GuestSessionResponse:
-    """Import a workspace zip and switch the current guest session to it."""
-    store = getattr(request.app.state, "workspace_store", None)
-    manager = getattr(request.app.state, "guest_session_manager", None)
-    if not store or not manager:
-        raise HTTPException(status_code=503, detail="Workspace service unavailable")
-
-    zip_bytes = await archive.read()
-    try:
-        workspace = store.import_zip(zip_bytes)
-    except ValueError as err:
-        raise HTTPException(status_code=400, detail=str(err)) from err
-
-    response.set_cookie(
-        manager.cookie_name,
-        manager.encode(workspace.id),
-        httponly=True,
-        secure=COOKIE_SECURE,
-        samesite=COOKIE_SAMESITE,
-        max_age=manager.cookie_max_age_seconds(),
-    )
-    return GuestSessionResponse(workspace_id=workspace.id, created=True)
-
-
-@app.get("/api/characters", response_model=CharactersListResponse)
-async def get_characters_api(request: Request) -> CharactersListResponse:
-    """Unversioned REST endpoint for characters list."""
-    defaults = (await get_characters()).characters
-    workspace_id = _get_optional_workspace_id(request)
-    workspace_chars: List[str] = []
-    store = getattr(request.app.state, "workspace_character_store", None)
-    if workspace_id and store:
-        try:
-            workspace_chars = store.list_ids(workspace_id)
-        except Exception:
-            workspace_chars = []
-    return CharactersListResponse(characters=sorted({*defaults, *workspace_chars}))
-
-
-@app.get("/api/characters/{character_id}", response_model=CharacterDetailResponse)
-async def get_character_detail_api(character_id: str, request: Request) -> CharacterDetailResponse:
-    """Unversioned REST endpoint for character detail."""
-    workspace_id = _get_optional_workspace_id(request)
-    store = getattr(request.app.state, "workspace_character_store", None)
-    if workspace_id and store:
-        try:
-            record = store.get(workspace_id, character_id)
-        except ValueError as err:
-            raise HTTPException(status_code=400, detail=str(err)) from err
-        if record:
-            return _workspace_record_to_character_detail(record)
-    return await get_character_detail(character_id)
-
-
-@app.post("/api/characters", response_model=CharacterDetailResponse, status_code=201)
-async def create_workspace_character(
-    request: Request,
-    response: Response,
-    payload: WorkspaceCharacterCreateRequest,
-    workspace_id: str = Depends(_require_workspace_id),
-) -> CharacterDetailResponse:
-    """Create a character inside the current guest workspace."""
-    store = getattr(request.app.state, "workspace_character_store", None)
-    if not store:
-        raise HTTPException(status_code=503, detail="Workspace service unavailable")
-
-    requested_id = payload.agent_id or _normalize_character_id(payload.name)
-    character_id = _normalize_character_id(requested_id)
-
-    record_payload: Dict[str, Any] = {
-        "name": payload.name.strip(),
-        "background_summary": payload.background_summary.strip(),
-        "personality_traits": payload.personality_traits.strip(),
-        "skills": payload.skills or {},
-        "relationships": payload.relationships or {},
-        "current_location": payload.current_location or "",
-        "inventory": payload.inventory or [],
-        "metadata": payload.metadata or {},
-        "structured_data": payload.structured_data or {},
-        "current_status": "active",
-        "narrative_context": "",
-    }
-
-    # Ensure uniqueness within workspace.
-    candidate = character_id
-    suffix = 2
-    while store.get(workspace_id, candidate):
-        candidate = f"{character_id}_{suffix}"
-        suffix += 1
-    character_id = candidate
-
-    try:
-        record = store.create(workspace_id, character_id, record_payload)
-    except FileExistsError as err:
-        raise HTTPException(status_code=409, detail=str(err)) from err
-    except ValueError as err:
-        raise HTTPException(status_code=400, detail=str(err)) from err
-
-    return _workspace_record_to_character_detail(record)
-
-
-@app.put("/api/characters/{character_id}", response_model=CharacterDetailResponse)
-async def update_workspace_character(
-    character_id: str,
-    request: Request,
-    response: Response,
-    updates: WorkspaceCharacterUpdateRequest,
-    workspace_id: str = Depends(_require_workspace_id),
-) -> CharacterDetailResponse:
-    """Update a character inside the current guest workspace."""
-    store = getattr(request.app.state, "workspace_character_store", None)
-    if not store:
-        raise HTTPException(status_code=503, detail="Workspace service unavailable")
-    try:
-        record = store.update(workspace_id, character_id, updates.dict(exclude_unset=True))
-    except FileNotFoundError as err:
-        raise HTTPException(status_code=404, detail=str(err)) from err
-    except ValueError as err:
-        raise HTTPException(status_code=400, detail=str(err)) from err
-    return _workspace_record_to_character_detail(record)
-
-
-@app.delete("/api/characters/{character_id}", status_code=204)
-async def delete_workspace_character(
-    character_id: str,
-    request: Request,
-    response: Response,
-    workspace_id: str = Depends(_require_workspace_id),
-) -> None:
-    """Delete a character from the current guest workspace."""
-    store = getattr(request.app.state, "workspace_character_store", None)
-    if not store:
-        raise HTTPException(status_code=503, detail="Workspace service unavailable")
-    try:
-        store.delete(workspace_id, character_id)
-    except ValueError as err:
-        raise HTTPException(status_code=400, detail=str(err)) from err
-    return None
-
-
-@app.get("/api/characters/{character_id}/enhanced")
-async def get_character_enhanced_api(character_id: str) -> Dict[str, Any]:
-    """Unversioned REST endpoint for enhanced character detail."""
-    return await get_character_enhanced(character_id)
-
-
-# ===================================================================
-# Real-time Events SSE Endpoint (Dashboard Live API Integration)
-# ===================================================================
-
-# Track active SSE connections for monitoring
-active_sse_connections = {"count": 0}
-
-# Global event queue for SSE - stores events from EventBus for all connected clients
-# Each client gets events from this shared queue via asyncio.Queue per client
-sse_event_queues: Dict[str, asyncio.Queue] = {}
-sse_event_id_counter = {"value": 0}
-
-def create_sse_event(
-    event_type: str,
-    title: str,
-    description: str,
-    severity: str = "low",
-    character_name: Optional[str] = None
-) -> Dict[str, Any]:
-    """Create a properly formatted SSE event payload."""
-    sse_event_id_counter["value"] += 1
-    event_id = sse_event_id_counter["value"]
-
-    event_data = {
-        "id": f"evt-{event_id}",
-        "type": event_type,
-        "title": title,
-        "description": description,
-        "timestamp": int(time.time() * 1000),
-        "severity": severity,
-    }
-
-    if character_name:
-        event_data["characterName"] = character_name
-
-    return event_data
-
-# Global event loop reference for thread-safe SSE broadcasting
-_main_loop: Optional[asyncio.AbstractEventLoop] = None
-
-@app.on_event("startup")
-async def startup_event():
-    global _main_loop
-    _main_loop = asyncio.get_running_loop()
-    logger.info("Captured main event loop for thread-safe operations")
-
-def _safe_sse_put(queue: asyncio.Queue, data: Dict[str, Any], client_id: str):
-    """Helper to safely put data into SSE queue from another thread."""
-    try:
-        queue.put_nowait(data)
-    except asyncio.QueueFull:
-        logger.warning(f"SSE queue full for client {client_id}, dropping event")
-
-def broadcast_sse_event(event_data: Dict[str, Any]):
-    """Broadcast an event to all connected SSE clients (thread-safe)."""
-    global _main_loop
-    
-    logger.info("Broadcasting SSE event to %d clients", len(sse_event_queues))
-
-    # If we don't have the loop yet (e.g. unit tests), try to get it or skip
-    if _main_loop is None:
-        try:
-            loop = asyncio.get_running_loop()
-        except RuntimeError:
-            logger.warning("No event loop available for SSE broadcast")
-            return
-    else:
-        loop = _main_loop
-
-    for client_id, queue in list(sse_event_queues.items()):
-        if loop.is_running():
-            loop.call_soon_threadsafe(_safe_sse_put, queue, event_data, client_id)
-
-
-
-# Initialize decision system now that broadcast_sse_event is available
-if DECISION_ROUTER_AVAILABLE:
-    try:
-        from src.decision import initialize_decision_system as _init_decision
-        _init_decision(
-            pause_controller=_decision_pause_controller,
-            decision_detector=_decision_detector,
-            negotiation_engine=_decision_negotiation_engine,
-            broadcast_sse_event=broadcast_sse_event,
-        )
-        logger.info("Decision system initialized with SSE broadcast capability")
-    except Exception as e:
-        logger.warning(f"Failed to initialize decision system: {e}")
-
-
-def setup_eventbus_sse_bridge(event_bus: EventBus):
-    """
-    Subscribe to EventBus events and broadcast them to SSE clients.
-    This bridges the internal event system to the dashboard real-time feed.
-    """
-    def on_character_event(*args, **kwargs):
-        """Handle character-related events from EventBus."""
-        character_name = kwargs.get("character_name") or kwargs.get("name") or "Unknown"
-        action = kwargs.get("action") or kwargs.get("event") or "action"
-        description = kwargs.get("description") or f"{character_name} performed {action}"
-
-        event_data = create_sse_event(
-            event_type="character",
-            title=f"Character: {character_name}",
-            description=description,
-            severity="low",
-            character_name=character_name
-        )
-        broadcast_sse_event(event_data)
-
-    def on_story_event(*args, **kwargs):
-        """Handle story progression events from EventBus."""
-        title = kwargs.get("title") or "Story Update"
-        description = kwargs.get("description") or kwargs.get("text") or "Narrative progressed"
-
-        event_data = create_sse_event(
-            event_type="story",
-            title=title,
-            description=description,
-            severity="medium"
-        )
-        broadcast_sse_event(event_data)
-
-    def on_system_event(*args, **kwargs):
-        """Handle system events from EventBus."""
-        title = kwargs.get("title") or "System"
-        description = kwargs.get("description") or kwargs.get("message") or "System event"
-        severity = kwargs.get("severity") or "low"
-
-        event_data = create_sse_event(
-            event_type="system",
-            title=title,
-            description=description,
-            severity=severity
-        )
-        broadcast_sse_event(event_data)
-
-    def on_interaction_event(*args, **kwargs):
-        """Handle interaction events from EventBus."""
-        participants = kwargs.get("participants") or []
-        description = kwargs.get("description") or "Interaction occurred"
-        character_name = participants[0] if participants else None
-
-        event_data = create_sse_event(
-            event_type="interaction",
-            title="Interaction",
-            description=description,
-            severity="low",
-            character_name=character_name
-        )
-        broadcast_sse_event(event_data)
-
-    # Subscribe to various event types
-    event_bus.subscribe("character_action", on_character_event)
-    event_bus.subscribe("character_update", on_character_event)
-    event_bus.subscribe("story_progression", on_story_event)
-    event_bus.subscribe("narrative_update", on_story_event)
-    event_bus.subscribe("system_alert", on_system_event)
-    event_bus.subscribe("system_status", on_system_event)
-    event_bus.subscribe("interaction", on_interaction_event)
-    event_bus.subscribe("character_interaction", on_interaction_event)
-
-    # Also subscribe to generic events
-    event_bus.subscribe("dashboard_event", lambda *args, **kwargs: broadcast_sse_event(
-        create_sse_event(
-            event_type=kwargs.get("type", "system"),
-            title=kwargs.get("title", "Event"),
-            description=kwargs.get("description", "Dashboard event"),
-            severity=kwargs.get("severity", "low"),
-            character_name=kwargs.get("character_name")
-        )
-    ))
-
-    logger.info("EventBus-SSE bridge configured with event subscriptions")
-
-async def event_generator(client_id: str):
-    """
-    Async generator yielding SSE-formatted events for real-time dashboard updates.
-
-    Streams events with format:
-    - retry: <milliseconds>
-    - id: <event-id>
-    - data: <json-payload>
-
-    Args:
-        client_id: Unique identifier for the connected client
-
-    Yields:
-        SSE-formatted event strings
-    """
-    # Create a queue for this client
-    client_queue: asyncio.Queue = asyncio.Queue(maxsize=100)
-    sse_event_queues[client_id] = client_queue
-
-    try:
-        # Send retry directive for client reconnection interval (3 seconds)
-        yield "retry: 3000\n\n"
-
-        logger.info(f"SSE client connected: {client_id}")
-        active_sse_connections["count"] += 1
-
-        # Send initial connection event
-        connect_event = create_sse_event(
-            event_type="system",
-            title="Connected",
-            description="Real-time event stream connected",
-            severity="low"
-        )
-        yield f"id: {connect_event['id']}\n"
-        yield f"data: {json.dumps(connect_event)}\n\n"
-
-        # Main event loop - wait for real events from the queue
-        while True:
-            try:
-                # Wait for events with timeout (sends heartbeat if no events)
-                try:
-                    event_data = await asyncio.wait_for(client_queue.get(), timeout=30.0)
-
-                    # Format as SSE message
-                    yield f"id: {event_data['id']}\n"
-                    yield f"data: {json.dumps(event_data)}\n\n"
-
-                except asyncio.TimeoutError:
-                    # Send heartbeat comment to keep connection alive
-                    yield ": heartbeat\n\n"
-
-            except asyncio.CancelledError:
-                # Client disconnected - clean up and exit gracefully
-                logger.info(f"SSE client disconnected: {client_id}")
-                break
-
-            except Exception as e:
-                # Internal error - send error event but continue streaming
-                logger.error(f"SSE event generation error for client {client_id}: {e}")
-
-                error_event = create_sse_event(
-                    event_type="system",
-                    title="Stream Error",
-                    description=f"Internal error: {str(e)}",
-                    severity="high"
-                )
-
-                yield f"id: {error_event['id']}\n"
-                yield f"data: {json.dumps(error_event)}\n\n"
-
-    except Exception as fatal_error:
-        # Fatal error - log and terminate
-        logger.error(f"Fatal SSE error for client {client_id}: {fatal_error}")
-        raise
-    finally:
-        # Clean up client queue
-        active_sse_connections["count"] -= 1
-        if client_id in sse_event_queues:
-            del sse_event_queues[client_id]
-        logger.info(f"SSE client {client_id} cleaned up")
-
-
-@app.get("/api/events/stream", tags=["Dashboard"])
-async def stream_events():
-    """
-    Server-Sent Events (SSE) endpoint for real-time dashboard events.
-
-    Streams continuous event updates with automatic reconnection support.
-    Events include: character actions, story progression, system alerts, interactions.
-
-    Returns:
-        StreamingResponse: text/event-stream with continuous event updates
-
-    Example:
-        ```javascript
-        const eventSource = new EventSource('/api/events/stream');
-        eventSource.onmessage = (event) => {
-            const data = JSON.parse(event.data);
-            console.log('Event:', data);
-        };
-        ```
-    """
-    # Generate unique client identifier for logging/monitoring
-    client_id = secrets.token_hex(8)
-
-    return StreamingResponse(
-        event_generator(client_id),
-        media_type="text/event-stream"
-    )
-
-
-class EmitEventRequest(BaseModel):
-    """Request model for emitting dashboard events."""
-    type: str = Field(default="system", description="Event type: character, story, system, interaction")
-    title: str = Field(description="Event title")
-    description: str = Field(description="Event description")
-    severity: str = Field(default="low", description="Event severity: low, medium, high")
-    character_name: Optional[str] = Field(default=None, description="Character name for character events")
-
-
-@app.post("/api/events/emit", tags=["Dashboard"])
-async def emit_dashboard_event(request: EmitEventRequest):
-    """
-    Emit a dashboard event to all connected SSE clients.
-
-    This endpoint allows triggering real-time events that will appear
-    in the dashboard's Real-Time Activity section.
-
-    Args:
-        request: Event details including type, title, description, severity
-
-    Returns:
-        Success confirmation with event ID
-    """
-    event_data = create_sse_event(
-        event_type=request.type,
-        title=request.title,
-        description=request.description,
-        severity=request.severity,
-        character_name=request.character_name
-    )
-
-    broadcast_sse_event(event_data)
-
-    return {
-        "success": True,
-        "message": "Event broadcast to all connected clients",
-        "event_id": event_data["id"],
-        "connected_clients": active_sse_connections["count"]
-    }
-
-
-@app.get("/api/events/stats", tags=["Dashboard"])
-async def get_sse_stats():
-    """Get SSE connection statistics."""
-    return {
-        "connected_clients": active_sse_connections["count"],
-        "total_events_sent": sse_event_id_counter["value"],
-        "active_queues": len(sse_event_queues)
-    }
-
-
-# Analytics metrics storage (could be persisted to database in production)
-analytics_metrics = {
-    "story_quality": 0.0,
-    "engagement": 0.0,
-    "coherence": 0.0,
-    "complexity": 0.0,
-    "data_points": 0,
-    "last_updated": None,
-}
-
-
-@app.get("/api/analytics/metrics", tags=["Dashboard"])
-async def get_analytics_metrics():
-    """
-    Get story analytics metrics for the dashboard.
-
-    Returns metrics like story quality, engagement, coherence, and complexity
-    based on current orchestration state and cache performance.
-    """
-    # Calculate real metrics based on system state
-    orchestration = await get_orchestration_status()
-    orch_data = orchestration.get("data", {}) if orchestration.get("success") else {}
-
-    # Get cache metrics for data points calculation
-    cache_metrics_raw = {}
-    try:
-        if chunk_cache:
-            # Try to get metrics if the method exists
-            if hasattr(chunk_cache, 'get_metrics'):
-                cache_metrics_raw = chunk_cache.get_metrics()
-            elif hasattr(chunk_cache, '_cache'):
-                # Fallback: construct metrics from internal state
-                cache_metrics_raw = {
-                    "cache_size": len(chunk_cache._cache) if hasattr(chunk_cache._cache, '__len__') else 0,
-                    "cache_semantic_hits": 0,
-                    "cache_exact_hits": 0,
-                }
-    except Exception:
-        pass  # Use empty dict if cache metrics unavailable
-
-    # Calculate metrics based on real system state
-    total_turns = orch_data.get("total_turns", 0)
-    current_turn = orch_data.get("current_turn", 0)
-    status = orch_data.get("status", "idle")
-
-    # Story quality: based on completed turns and processing success
-    completed_steps = sum(1 for step in orch_data.get("steps", []) if step.get("status") == "completed")
-    total_steps = len(orch_data.get("steps", []))
-    story_quality = (completed_steps / total_steps * 10) if total_steps > 0 else 8.0
-
-    # Engagement: based on SSE connections and event activity
-    active_clients = active_sse_connections.get("count", 0)
-    total_events = sse_event_id_counter.get("value", 0)
-    engagement = min(100, 70 + active_clients * 10 + min(total_events, 30))
-
-    # Coherence: based on cache hit rate (semantic consistency)
-    cache_hits = cache_metrics_raw.get("cache_semantic_hits", 0) + cache_metrics_raw.get("cache_exact_hits", 0)
-    cache_size = cache_metrics_raw.get("cache_size", 0)
-    coherence = min(100, 85 + (cache_hits / max(1, cache_size)) * 15) if cache_size > 0 else 90
-
-    # Complexity: based on number of characters and active steps
-    complexity = min(10, 6 + (total_steps * 0.3) + (current_turn * 0.1))
-
-    # Data points: sum of cached items and events
-    data_points = cache_size + total_events + total_turns
-
-    return {
-        "success": True,
-        "data": {
-            "story_quality": round(story_quality, 1),
-            "engagement": round(engagement, 0),
-            "coherence": round(coherence, 0),
-            "complexity": round(complexity, 1),
-            "data_points": data_points,
-            "metrics_tracked": 5,
-            "status": status,
-            "last_updated": datetime.now(UTC).isoformat(),
-        }
-    }
-
-
-@app.post("/campaigns", response_model=CampaignCreationResponse)
-async def create_campaign(request: CampaignCreationRequest) -> CampaignCreationResponse:
-    """Creates a new campaign."""
-    try:
-        campaign_id = f"campaign_{uuid.uuid4().hex[:8]}"
-        campaigns_dir = "campaigns"
-        os.makedirs(campaigns_dir, exist_ok=True)
-
-        campaign_file = os.path.join(campaigns_dir, f"{campaign_id}.json")
-        campaign_data = {
-            "id": campaign_id,
-            "name": request.name,
-            "description": request.description or "",
-            "participants": request.participants,
-            "created_at": datetime.now().isoformat(),
-            "status": "created",
-        }
-
-        with open(campaign_file, "w") as f:
-            json.dump(campaign_data, f, indent=2)
-
-        logger.info("Created campaign %s", campaign_id)
-        return CampaignCreationResponse(
-            campaign_id=campaign_id,
-            name=request.name,
-            status="created",
-            created_at=campaign_data["created_at"],
-        )
-
-    except Exception as e:
-        logger.error(f"Error creating campaign: {e}", exc_info=True)
-        raise HTTPException(
-            status_code=500, detail=f"Failed to create campaign: {str(e)}"
-        )
-
-
-# ===================================================================
-# Authentication Endpoints (Security Improvements - SEC-001)
-# ===================================================================
-
-@app.post("/api/auth/login", response_model=AuthResponse, tags=["Authentication"])
-async def login(credentials: LoginRequest, response: Response):
-    """
-    Authenticate user and set httpOnly cookies for token storage.
-
-    This endpoint implements SEC-001 security improvements:
-    - Tokens stored in httpOnly cookies (XSS protection)
-    - Secure flag for HTTPS-only transmission
-    - SameSite=Lax for CSRF protection
-    - Short-lived access tokens (15 min)
-    - Long-lived refresh tokens (30 days)
-
-    The response also includes tokens in the body for backward compatibility
-    during the migration period.
-
-    Args:
-        credentials: User login credentials (email, password, remember_me)
-        response: FastAPI Response object for setting cookies
-
-    Returns:
-        AuthResponse with tokens and user information
-    """
-    try:
-        # TODO: Replace with actual authentication logic
-        # For now, this is a mock implementation for demonstration
-
-        # Validate credentials (mock implementation)
-        if not credentials.email or not credentials.password:
-            raise HTTPException(
-                status_code=400,
-                detail="Email and password are required"
-            )
-
-        # Mock user data - replace with actual database lookup
-        user_data = {
-            "id": str(uuid.uuid4()),
-            "email": credentials.email,
-            "name": credentials.email.split("@")[0],
-            "role": "user",
-            "created_at": datetime.now(UTC).isoformat(),
-            "updated_at": datetime.now(UTC).isoformat(),
-        }
-
-        # Calculate token expiry
-        access_token_expires = datetime.now(UTC) + timedelta(minutes=JWT_ACCESS_TOKEN_EXPIRE_MINUTES)
-        refresh_token_expires = datetime.now(UTC) + timedelta(days=30)
-
-        # Create JWT tokens
-        access_token_payload = {
-            "user_id": user_data["id"],
-            "email": user_data["email"],
-            "exp": access_token_expires,
-            "iat": datetime.now(UTC),
-            "type": "access"
-        }
-
-        refresh_token_payload = {
-            "user_id": user_data["id"],
-            "email": user_data["email"],
-            "exp": refresh_token_expires,
-            "iat": datetime.now(UTC),
-            "type": "refresh"
-        }
-
-        access_token = jwt.encode(access_token_payload, JWT_SECRET_KEY, algorithm=JWT_ALGORITHM)
-        refresh_token = jwt.encode(refresh_token_payload, JWT_SECRET_KEY, algorithm=JWT_ALGORITHM)
-
-        # Set httpOnly cookies for tokens
-        cookie_max_age = REFRESH_COOKIE_MAX_AGE if credentials.remember_me else COOKIE_MAX_AGE
-
-        response.set_cookie(
-            key=COOKIE_NAME,
-            value=access_token,
-            httponly=COOKIE_HTTPONLY,
-            secure=COOKIE_SECURE,
-            samesite=COOKIE_SAMESITE,
-            max_age=COOKIE_MAX_AGE
-        )
-
-        response.set_cookie(
-            key=REFRESH_COOKIE_NAME,
-            value=refresh_token,
-            httponly=COOKIE_HTTPONLY,
-            secure=COOKIE_SECURE,
-            samesite=COOKIE_SAMESITE,
-            max_age=cookie_max_age
-        )
-
-        logger.info("User login successful (cookies set)")
-
-        # Return tokens in response body for backward compatibility
-        return AuthResponse(
-            access_token=access_token,
-            refresh_token=refresh_token,
-            token_type="Bearer",
-            expires_in=JWT_ACCESS_TOKEN_EXPIRE_MINUTES * 60,
-            user=user_data
-        )
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"Login failed: {e}", exc_info=True)
-        raise HTTPException(
-            status_code=500,
-            detail=f"Login failed: {str(e)}"
-        )
-
-
-@app.post("/api/auth/refresh", response_model=AuthResponse, tags=["Authentication"])
-async def refresh_token(request: RefreshTokenRequest, response: Response, req: Request):
-    """
-    Refresh access token using refresh token.
-
-    Accepts refresh token from:
-    1. httpOnly cookie (preferred)
-    2. Request body (for backward compatibility)
-
-    Returns new access token and sets httpOnly cookie.
-
-    Args:
-        request: Refresh token request (optional, for backward compatibility)
-        response: FastAPI Response object for setting cookies
-        req: FastAPI Request object for reading cookies
-
-    Returns:
-        AuthResponse with new tokens
-    """
-    try:
-        # Try to get refresh token from cookie first, then from body
-        refresh_token_value = req.cookies.get(REFRESH_COOKIE_NAME) or request.refresh_token
-
-        if not refresh_token_value:
-            raise HTTPException(
-                status_code=401,
-                detail="No refresh token provided"
-            )
-
-        # Validate and decode refresh token
-        try:
-            payload = jwt.decode(refresh_token_value, JWT_SECRET_KEY, algorithms=[JWT_ALGORITHM])
-
-            if payload.get("type") != "refresh":
-                raise HTTPException(
-                    status_code=401,
-                    detail="Invalid token type"
-                )
-
-            user_id = payload.get("user_id")
-            email = payload.get("email")
-
-            if not user_id or not email:
-                raise HTTPException(
-                    status_code=401,
-                    detail="Invalid token payload"
-                )
-
-        except jwt.ExpiredSignatureError:
-            raise HTTPException(
-                status_code=401,
-                detail="Refresh token has expired"
-            )
-        except jwt.InvalidTokenError as e:
-            raise HTTPException(
-                status_code=401,
-                detail=f"Invalid refresh token: {str(e)}"
-            )
-
-        # TODO: Verify user still exists and is active
-        user_data = {
-            "id": user_id,
-            "email": email,
-            "name": email.split("@")[0],
-            "role": "user",
-            "created_at": datetime.now(UTC).isoformat(),
-            "updated_at": datetime.now(UTC).isoformat(),
-        }
-
-        # Create new access token
-        access_token_expires = datetime.now(UTC) + timedelta(minutes=JWT_ACCESS_TOKEN_EXPIRE_MINUTES)
-        access_token_payload = {
-            "user_id": user_id,
-            "email": email,
-            "exp": access_token_expires,
-            "iat": datetime.now(UTC),
-            "type": "access"
-        }
-
-        access_token = jwt.encode(access_token_payload, JWT_SECRET_KEY, algorithm=JWT_ALGORITHM)
-
-        # Set new access token cookie
-        response.set_cookie(
-            key=COOKIE_NAME,
-            value=access_token,
-            httponly=COOKIE_HTTPONLY,
-            secure=COOKIE_SECURE,
-            samesite=COOKIE_SAMESITE,
-            max_age=COOKIE_MAX_AGE
-        )
-
-        logger.info(f"Token refreshed for user: {email}")
-
-        return AuthResponse(
-            access_token=access_token,
-            refresh_token=refresh_token_value,
-            token_type="Bearer",
-            expires_in=JWT_ACCESS_TOKEN_EXPIRE_MINUTES * 60,
-            user=user_data
-        )
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"Token refresh failed: {e}", exc_info=True)
-        raise HTTPException(
-            status_code=500,
-            detail=f"Token refresh failed: {str(e)}"
-        )
-
-
-@app.get("/api/auth/csrf-token", response_model=CSRFTokenResponse, tags=["Authentication"])
-async def get_csrf_token(response: Response):
-    """
-    Generate and return CSRF token for state-changing requests.
-
-    The CSRF token is set as a non-httpOnly cookie so JavaScript can read it
-    and include it in request headers. This provides CSRF protection for
-    state-changing operations (POST, PUT, DELETE, PATCH).
-
-    Usage:
-    1. Call this endpoint to get a CSRF token
-    2. Include the token in X-CSRF-Token header for mutations
-    3. Backend validates the token matches the cookie value
-
-    Returns:
-        CSRFTokenResponse with the generated CSRF token
-    """
-    try:
-        # Generate cryptographically secure random token
-        csrf_token = secrets.token_urlsafe(32)
-
-        # Set CSRF token as non-httpOnly cookie (JS needs to read it)
-        response.set_cookie(
-            key=CSRF_COOKIE_NAME,
-            value=csrf_token,
-            httponly=False,  # JavaScript must be able to read this
-            secure=COOKIE_SECURE,
-            samesite="strict",  # Strict for CSRF tokens
-            max_age=CSRF_TOKEN_MAX_AGE
-        )
-
-        logger.debug("CSRF token generated and set")
-
-        return CSRFTokenResponse(csrf_token=csrf_token)
-
-    except Exception as e:
-        logger.error(f"CSRF token generation failed: {e}", exc_info=True)
-        raise HTTPException(
-            status_code=500,
-            detail=f"Failed to generate CSRF token: {str(e)}"
-        )
-
-
-class LogoutRequest(BaseModel):
-    """Request model for logout."""
-    access_token: Optional[str] = Field(None, description="Optional access token to invalidate")
-
-
-class LogoutResponse(BaseModel):
-    """Response model for logout."""
-    success: bool
-    message: str
-
-
-class TokenValidationResponse(BaseModel):
-    """Response model for token validation."""
-    valid: bool
-    expires_at: Optional[int] = Field(None, description="Token expiry timestamp in milliseconds")
-    user_id: Optional[str] = None
-    error: Optional[str] = None
-
-
-@app.post("/api/auth/logout", response_model=LogoutResponse, tags=["Authentication"])
-async def logout(request: Request, response: Response, body: Optional[LogoutRequest] = None):
-    """
-    Logout endpoint to invalidate user session and clear httpOnly cookies.
-
-    This endpoint implements SEC-001 security improvements:
-    - Clears all authentication cookies (access_token, refresh_token)
-    - Logs the logout event for audit trail
-    - Always returns success to ensure frontend can reliably clear state
-
-    The endpoint clears httpOnly cookies by setting them to expire immediately.
-
-    In a production environment with a session store, this would also:
-    - Add tokens to a Redis blacklist with TTL matching token expiry
-    - Mark session as inactive in database
-
-    Args:
-        request: FastAPI Request object for logging
-        response: FastAPI Response object for clearing cookies
-        body: Optional logout request body (for backward compatibility)
-
-    Returns:
-        LogoutResponse with success status
-    """
-    try:
-        # Extract token from Authorization header or request body for logging
-        auth_header = request.headers.get("Authorization", "")
-        token = None
-
-        if auth_header.startswith("Bearer "):
-            token = auth_header[7:]
-        elif body and body.access_token:
-            token = body.access_token
-
-        # Log the logout event
-        client_ip = request.client.host if request.client else "unknown"
-        user_agent = request.headers.get("user-agent", "unknown")
-
-        # Mask token for logging (show first 10 chars only)
-        masked_token = f"{token[:10]}..." if token and len(token) > 10 else "no-token"
-
-        logger.info(
-            f"Logout event: token={masked_token}, ip={client_ip}, ua={user_agent[:50]}"
-        )
-
-        # Clear httpOnly cookies by setting max_age=0
-        response.delete_cookie(
-            key=COOKIE_NAME,
-            secure=COOKIE_SECURE,
-            samesite=COOKIE_SAMESITE
-        )
-
-        response.delete_cookie(
-            key=REFRESH_COOKIE_NAME,
-            secure=COOKIE_SECURE,
-            samesite=COOKIE_SAMESITE
-        )
-
-        response.delete_cookie(
-            key=CSRF_COOKIE_NAME,
-            secure=COOKIE_SECURE,
-            samesite="strict"
-        )
-
-        logger.info("Authentication cookies cleared successfully")
-
-        # TODO: [SEC-002] When implementing a token blacklist or session store,
-        # add token invalidation here:
-        # - Add token to a Redis blacklist with TTL matching token expiry
-        # - Or mark session as inactive in database
-
-        return LogoutResponse(
-            success=True,
-            message="Logout successful"
-        )
-
-    except Exception as e:
-        # Always return success for logout to ensure frontend can clear state
-        logger.warning(f"Logout encountered an error (still returning success): {e}")
-        return LogoutResponse(
-            success=True,
-            message="Logout successful"
-        )
-
-
-@app.get("/api/auth/validate", response_model=TokenValidationResponse, tags=["Authentication"])
-async def validate_token(request: Request):
-    """
-    Validate the current access token.
-
-    This endpoint checks if the token in the Authorization header is valid.
-    Returns token status and expiry information.
-
-    Returns:
-        200 with valid=true if token is valid
-        401 with valid=false if token is invalid or expired
-    """
-    try:
-        auth_header = request.headers.get("Authorization", "")
-
-        if not auth_header.startswith("Bearer "):
-            return JSONResponse(
-                status_code=401,
-                content=TokenValidationResponse(
-                    valid=False,
-                    error="Missing or invalid Authorization header"
-                ).model_dump()
-            )
-
-        token = auth_header[7:]
-
-        if not token:
-            return JSONResponse(
-                status_code=401,
-                content=TokenValidationResponse(
-                    valid=False,
-                    error="No token provided"
-                ).model_dump()
-            )
-
-        # Try to decode and validate the token
-        try:
-            # Use the module-level JWT_SECRET_KEY constant for consistency
-            secret_key = JWT_SECRET_KEY
-
-            # Decode the token
-            payload = jwt.decode(token, secret_key, algorithms=["HS256"])
-
-            # Extract expiry and user info
-            exp = payload.get("exp")
-            user_id = payload.get("user_id") or payload.get("sub")
-
-            # Check if token is expired
-            if exp:
-                from datetime import timezone
-                exp_datetime = datetime.fromtimestamp(exp, tz=timezone.utc)
-                now = datetime.now(timezone.utc)
-
-                if now > exp_datetime:
-                    return JSONResponse(
-                        status_code=401,
-                        content=TokenValidationResponse(
-                            valid=False,
-                            error="Token has expired"
-                        ).model_dump()
-                    )
-
-                # Return valid token with expiry in milliseconds
-                expires_at_ms = int(exp * 1000)
-
-                return TokenValidationResponse(
-                    valid=True,
-                    expires_at=expires_at_ms,
-                    user_id=user_id
-                )
-
-            # Token is valid but no expiry (shouldn't happen with proper tokens)
-            return TokenValidationResponse(
-                valid=True,
-                user_id=user_id
-            )
-
-        except jwt.ExpiredSignatureError:
-            return JSONResponse(
-                status_code=401,
-                content=TokenValidationResponse(
-                    valid=False,
-                    error="Token has expired"
-                ).model_dump()
-            )
-        except jwt.InvalidTokenError as e:
-            return JSONResponse(
-                status_code=401,
-                content=TokenValidationResponse(
-                    valid=False,
-                    error=f"Invalid token: {str(e)}"
-                ).model_dump()
-            )
-
-    except Exception as e:
-        logger.error(f"Token validation error: {e}")
-        return JSONResponse(
-            status_code=401,
-            content=TokenValidationResponse(
-                valid=False,
-                error="Token validation failed"
-            ).model_dump()
-        )
-
-
-# Include World Context Router (registered late to avoid shadowing static /api/* endpoints)
-if WORLD_ROUTER_AVAILABLE:
-    app.include_router(world_router, prefix="/api")
-    logger.info("World context router included with prefix /api")
-else:
-    logger.warning(
-        "World context router not available - endpoints will not be accessible"
-    )
-
-
-def run_server(host: str = "127.0.0.1", port: int = 8000, debug: bool = False):
-    """Runs the FastAPI server."""
+def run_server(host: str = "127.0.0.1", port: int = 8000, debug: bool = False) -> None:
+    """Run the FastAPI server via Uvicorn."""
     uvicorn.run("api_server:app", host=host, port=port, reload=debug, log_level="info")
 
 
@@ -2233,3 +53,4 @@ if __name__ == "__main__":
     port = int(os.getenv("API_PORT", "8000"))
     debug_flag = os.getenv("API_DEBUG", "1")
     run_server(host=host, port=port, debug=debug_flag not in {"0", "false", "False"})
+
