+++ src/api/routers/__init__.py
@@ -1,0 +1,2 @@
+"""FastAPI routers grouped by domain."""
+
