+++ src/api/services/__init__.py
@@ -1,0 +1,2 @@
+"""Domain services for the API layer."""
+
